Make watch file work
