+++ lib/assetTypeMap.js
