+++ lib/html/ejs.js
@@ -1,6 +1,5 @@
 'use strict';
 
-var Bluebird = require('bluebird');
 var resolve = require('resolve');
 
 module.exports = function renderEjs(code) {
