+++ lib/util/file.js
@@ -20,7 +20,7 @@ function watch(pattern) {
     });
 
     watcher.on('all', function(event, path) {
-      observer.onNext([event, path]);
+      observer.onNext(path);
     });
 
     return function () {
