+++ lib/util/identity-async.js
@@ -1,7 +1,5 @@
 'use strict';
 
-var Bluebird = require('bluebird');
-
 module.exports = function identityAsync(input) {
-  return Bluebird.resolve(input);
+  return input;
 };
