+++ lib/util/playground.js
@@ -7,10 +7,9 @@ var __ = R.__;
 var pipe = R.pipe;
 var defaultTo = R.defaultTo;
 var prop = R.prop;
-var assetTypeMap = require('../asset-type-map');
+var assetTypeMap = require('../assetTypeMap');
 var Functional = require('./functional');
 var later = Functional.later;
-var noop = Functional.noop;
 var File = require('./file');
 var readFile = File.readFile;
 var writeFile = File.writeFile;
@@ -57,8 +56,32 @@ function createAssetFile(playground, name, tmplFactory) {
   )(playground);
 }
 
+function parsePlaygroundFile(file) {
+  return readFile(file)
+    .map(function (content) {
+      return parse(getExt(file), content);
+    });
+}
+
+/**
+ * Get supported extensions for specified asset type
+ * @param  {string}   type Could be 'html', 'js' or 'css'
+ * @return {function}      Accept playground object as argument, return array
+ *                         of extension strings
+ */
+function extensionsForAsset(type) {
+  return pipe(
+    R.path([type, 'preprocessor']),
+    defaultTo(type),
+    prop(__, assetTypeMap[type].preProcessors),
+    prop('extensions')
+  );
+};
+
 module.exports = {
   getExt: getExt,
   parse: parse,
   createAssetFile: createAssetFile,
+  parsePlaygroundFile: parsePlaygroundFile,
+  extensionsForAsset: extensionsForAsset,
 };
