+++ lib/web-playground.js
@@ -1,103 +1,8 @@
 'use strict';
 
-var express = require('express');
-var chalk = require('chalk');
-var Observable = require('rx').Observable;
-var just = Observable.just;
-var combineLatest = Observable.combineLatest;
-var R = require('ramda');
-var pipe = R.pipe;
-var of = R.of;
-var append = R.append;
-var merge = R.merge;
-var Functional = require('./util/functional');
-var later = Functional.later;
-var Log = require('../lib/util/log');
-var info = Log.info;
-var logCreateFile = later(Log.logCreateFile);
-var File = require('./util/file');
-var readFile = File.readFile;
-var writeFile = File.writeFile;
-var watch = File.watch;
-var Playground = require('./util/playground');
-var parse = Playground.parse;
-var getExt = Playground.getExt;
-var createAssetFile = Playground.createAssetFile;
-var Observables = require('./util/observables');
-var tryInOrder = Observables.tryInOrder;
-var NoMoreFallbacksError = Observables.NoMoreFallbacksError;
-var CONSTANT = require('../lib/CONSTANT');
-var PLAYGROUND_FILES = CONSTANT.PLAYGROUND_FILES;
-var DEFAULT_PLAYGROUND = CONSTANT.DEFAULT_PLAYGROUND;
-var TMPL_PLAYGROUND = CONSTANT.TMPL_PLAYGROUND;
-var TMPL_HTML = CONSTANT.TMPL_HTML;
-var DEFAULT_OPTS = CONSTANT.DEFAULT_OPTS;
-
-// var attachData  = require('./middleware/attach-data');
-// var serveAssets = require('./middleware/serve-assets');
-
-function start(_opts) {
-  var opts = merge({}, DEFAULT_OPTS, _opts);
-
-  watch('{playground,html,css,js}.*')
-    .subscribeOnNext(console.log);
-
-  // var app = express();
-  // // app.use(attachData);
-  // // app.use(serveAssets);
-  // app.listen(opts.port);
-
-  // var bs = require('browser-sync').create();
-  // bs.watch().on('change', bs.reload);
-  // bs.init({ proxy: 'http://localhost:' + opts.port });
-};
+var preStart = require('./preStart');
+var start = require('./start');
 
 module.exports = function () {
-  var hasCreatedFromTmpl = false;
-
-  tryInOrder(PLAYGROUND_FILES, readFile)
-    .catch(function (err) {
-      if (!(err instanceof NoMoreFallbacksError)) throw err;
-      info('Did not find playground file, creating one...');
-      hasCreatedFromTmpl = true;
-      return readFile(TMPL_PLAYGROUND)
-        .flatMap(function (content) {
-          return writeFile(DEFAULT_PLAYGROUND, content)
-            .doOnNext(logCreateFile(DEFAULT_PLAYGROUND))
-            .flatMap(function () {
-              return just(content);
-            });
-        })
-        .map(pipe(of, append(DEFAULT_PLAYGROUND)));
-    })
-    .map(function (results) {
-      var content = results[0];
-      var name = results[1];
-      return parse(getExt(name), content);
-    })
-    .flatMap(function (playground) {
-      var htmlObs = createAssetFile(playground, 'html', function (ext) {
-        hasCreatedFromTmpl = true;
-        return ext === 'html' ? readFile(TMPL_HTML) : just('');
-      });
-
-      var jsObs = createAssetFile(playground, 'js', function (ext) {
-        hasCreatedFromTmpl = true;
-        return just('');
-      });
-
-      var cssObs = createAssetFile(playground, 'css', function (ext) {
-        hasCreatedFromTmpl = true;
-        return just('');
-      });
-
-      return combineLatest(htmlObs, jsObs, cssObs);
-    })
-    .subscribeOnCompleted(function () {
-      if (hasCreatedFromTmpl) {
-        info('Run ' + chalk.green('wpg') + ' again to start live-reload server');
-        return;
-      }
-      start();
-    });
+  preStart(start);
 };
